Add files via upload
This is a basic Python project that implements a number guessing game. The user is prompted to enter a maximum number, and the program generates a random number within that range. The user then tries to guess the number, receiving feedback on whether their guess is too high or too low, until they guess correctly. The number of attempts is displayed at the end.